sync: DZ pre 20240106

+++ src/hardware/Nest.cpp
@@ -644,8 +644,8 @@ void CNest::SetSetpoint(const int idx, const float temp)
 		if (!Login())
 			return;
 	}
-	size_t iThermostat = (idx - 1) / 3;
-	if (iThermostat > m_thermostats.size())
+	int iThermostat = (idx - 1) / 3;
+	if (iThermostat > (int)m_thermostats.size())
 		return;
 
 	std::vector<std::string> ExtraHeaders;
@@ -690,8 +690,8 @@ bool CNest::SetAway(const unsigned char Idx, const bool bIsAway)
 			return false;
 	}
 
-	size_t iThermostat = (Idx - 3) / 3;
-	if (iThermostat > m_thermostats.size())
+	int iThermostat = (Idx - 3) / 3;
+	if (iThermostat > (int)m_thermostats.size())
 		return false;
 
 	std::vector<std::string> ExtraHeaders;
@@ -730,8 +730,8 @@ bool CNest::SetManualEcoMode(const unsigned char Idx, const bool bIsManualEcoMod
 			return false;
 	}
 
-	size_t iThermostat = (Idx - 4) / 3;
-	if (iThermostat > m_thermostats.size())
+	int iThermostat = (Idx - 4) / 3;
+	if (iThermostat > (int)m_thermostats.size())
 		return false;
 
 	std::vector<std::string> ExtraHeaders;
